Polish antismash summary script
